sdks/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "o2-sdk"
version = "0.1.0"
description = "Python SDK for the O2 Exchange — a fully on-chain order book DEX on the Fuel Network"
readme = "AGENTS.md"
requires-python = ">=3.10"
license = "Apache-2.0"
dependencies = [
    "coincurve>=20.0.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "pycryptodome>=3.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.2",
    "ruff>=0.15.0",
    "mypy>=1.19.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/o2_sdk"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "module"
asyncio_default_test_loop_scope = "module"
markers = [
    "integration: marks tests that require network access (deselect with '-m \"not integration\"')",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["o2_sdk"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true
